git-finish: Dont automatically delete upstream
Automatically deleting the upstream causes CI systems
to fails when they try to fetch the freshly pushed branch,
which is not so nice.

Either we
1. delete the old branch without pushing, which causes
   Github to display "Closed" rather than "Merged", or
2. push without deleting, which causes an unnecessary
   CI build and leaves a dangling branch on Github.

1 is bad because it makes it a bit hard to find "abandoned" PRs as
opposed to properly merged PRs, but 2 is bad because it doesn't do
proper cleanup and causes unnecessary builds, which costs time.

I personally prefer 1 over 2, but should revisit this in the future.